SpreadsheetFormatterContextTesting2 extracted SpreadsheetFormatterContextTesting
- SpreadsheetFormatterContextTesting all methods require a SpreadsheetFormatterContext instance as the first parameter.
